Refactor: enum, struct 한 파일로 정리
InteractManager의 InteractWithPet 부분 중복 삭제 후,
PetBase의 interact~~ 함수로 통합
