+++ Assets/DongHyuk/Scripts/HandAttacher4Test.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections;
 using System.Collections.Generic;
+using EnumTypes;
 using UnityEngine;
 using Logger = ZipsAR.Logger;
