+++ Assets/TaeHyeon/Scripts/Gifts/GiftBoxSpawner.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections;
 using System.Collections.Generic;
+using EnumTypes;
 using UnityEngine;
 using Random = UnityEngine.Random;
