+++ Assets/TaeHyeon/Scripts/Pets/PetCollisionHandler.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections;
 using System.Collections.Generic;
+using EnumTypes;
 using UnityEngine;
 using UnityEngine.Serialization;
 using Logger = ZipsAR.Logger;
