+++ Assets/TaeHyeon/Scripts/Pets/PetObject.cs
@@ -1,5 +1,6 @@
 using System.Collections;
 using System.Collections.Generic;
+using EnumTypes;
 using UnityEngine;
 using UnityEngine.Serialization;
